Fixed a validation issue by removing the region in the aws_s3_bucket resource - see reason below --- Indeed, as @rice7646 noted this is the result of the completion of a deprecation cycle for that argument, which is now always forced to match the region configured in the provider. There’s more information on this in the [AWS provider v3 upgrade](https://registry.terraform.io/providers/hashicorp/aws/latest/docs/guides/version-3-upgrade#resource-aws_s3_bucket).
This problem is therefore not really specific to re-usable modules, but is rather just a major version
upgrade step for the AWS provider. If your intent is to write a module that could potentially be
compatible with AWS provider v2 and v3, I think it would suffice to remove the region argument and
let the provider = aws.alternative_region argument indirectly set your alternative region.

I believe the region argument was optional in AWS provider v2, as part of its deprecation process,
so a module that omits it should be compatible with both versions.